Move ArrayHelper::index tests to separately class.
